requirements.txt
Django==4.2.29
djangorestframework==3.16.1
dj-database-url==2.2.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.11.0
